[K/N] Fix general case of kt49827
When loading default libraries, the code assumed that all subdirectories
of <kotlin native home>/klib/*/ corresponded to valid libraries. The bug
report only considers the case of .DS_STORE, which is generated by
Finder on MacOS, but this case has already been accounted for in the
code.

The proposed fix downgrades a library parse errors (in the case of "try
to load all subfolders") down to a warning.  This should still provide
the programmer with enough information, in case the given folder turned
out to be a real library that cannot be parsed.